backend/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
mariadb==1.0.10
mysqlclient==2.1.0
numpy==1.22.3
pandas==1.4.1
Pillow==9.0.1
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
